Threads to be executed in the PayPal server
Three type of threads will at most be executed in the PayPal server.
- ServerListener is the main thread. It's mission is to accept new
clients and create new threads to satisfy their needs. Those thread
types can be the next two points.
- Registrator is a thread for the clients to be registererd. Clients
send data to this thread, and when the registration is finished the
connection is closed. If, after registering, the client wants to make a
payment, a new connection must be created.
- Payer is a type of thread to make payments. Clients are supposed to be
registered before (otherwise, no payment will be made). As it happens
with Registrator, clients send the necessary data to the Payer, and when
the payment is finished (doesn't need to be successfully completed), the
connection is closed.