DAO,VIEW,CONTROLLER,FACADE changes
I have added the view of eddit that is a copy of the one iñigo did , is the register but i will make some changes to make it an user eddit window. i have added the edit method in the controller and remote facade and also the DAO method.