Localization and DTO
I have added the localization for the messages on the eddit user. Also i have added all the data from user to the DTO, this is what i thought to be able to see all the data on the GUI but we clould change it if you think is not the best way of handlelinh this. This could be a problem because Guest and Admin has different data.