RequestedChanges(Read description Please)
I ahve changes what was requested. I also added a localization that was not added to the file. The thing about the dispose method, i have added to the class but how should i use it apart from adding the methods to the view. Every time i want to dispose a frame should i call this method instead of doing frame.dispose ???