Check Changed
I havent been able to get users from inside the reveiw. So i have changed the way the check to see if user has alredy posted a review about that hotel works. Now the ReveiwID is  mix of the userID that writtes and the hotelID the hotel the review is for. In this way if a reviewID is the same it means user alredy has a review on that hotel.